Ledger SDK V2

+++ bolos-impl/src/crypto/stark.rs
@@ -1,5 +1,5 @@
 /*******************************************************************************
-*   (c) 2021 Zondax GmbH
+*   (c) 2022 Zondax AG
 *
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
@@ -275,7 +275,7 @@ mod bindings {
         let mut sig_len = match crv.domain_length() {
             Some(n) => 6 + 2 * (n + 1),
             None => sig_out.len(),
-        } as u32;
+        };
 
         let mut info = 0;
